# docs-site/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tiao API",
    "version": "1.0.0",
    "description": "REST API and WebSocket protocol for Tiao, the open-source multiplayer board game platform."
  },
  "servers": [
    {
      "url": "http://localhost:5005/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/api/player/guest": {
      "post": {
        "summary": "Create a guest session",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/player/signup": {
      "post": {
        "summary": "Create a new account",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (missing fields, short password, short username, invalid email)"
          },
          "409": {
            "description": "Email or username already taken"
          },
          "503": {
            "description": "Account signup unavailable"
          }
        }
      }
    },
    "/api/player/login": {
      "post": {
        "summary": "Log in to an existing account",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier",
                  "password"
                ],
                "properties": {
                  "identifier": {
                    "type": "string",
                    "description": "Username or email address"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing identifier or password"
          },
          "401": {
            "description": "Account not found or incorrect password"
          },
          "503": {
            "description": "Account login unavailable"
          }
        }
      }
    },
    "/api/player/logout": {
      "post": {
        "summary": "Destroy the current session",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Session destroyed"
          }
        }
      }
    },
    "/api/player/me": {
      "get": {
        "summary": "Get the current authenticated player",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current player identity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "player": {
                      "$ref": "#/components/schemas/PlayerIdentity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated or session no longer valid"
          }
        }
      }
    },
    "/api/player/profile": {
      "get": {
        "summary": "Get the current account profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "profilePicture": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Only account players can access profiles"
          },
          "404": {
            "description": "Account not found"
          },
          "503": {
            "description": "Account features unavailable"
          }
        }
      },
      "put": {
        "summary": "Update the current account profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 3
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auth": {
                      "$ref": "#/components/schemas/AuthResponse"
                    },
                    "profile": {
                      "type": "object",
                      "properties": {
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "profilePicture": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error (no fields provided, short display name, short password, invalid email)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Only account players can edit profiles"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Username or email already taken"
          },
          "503": {
            "description": "Account features unavailable"
          }
        }
      }
    },
    "/api/player/profile-picture": {
      "post": {
        "summary": "Upload a profile picture",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "profilePicture"
                ],
                "properties": {
                  "profilePicture": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile picture uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auth": {
                      "$ref": "#/components/schemas/AuthResponse"
                    },
                    "profile": {
                      "type": "object",
                      "properties": {
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "profilePicture": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No image file provided"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Only account players can upload profile pictures"
          },
          "404": {
            "description": "Account not found"
          },
          "500": {
            "description": "Upload failed"
          },
          "503": {
            "description": "Account features unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PlayerIdentity": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "guest",
              "account"
            ]
          },
          "email": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "player": {
            "$ref": "#/components/schemas/PlayerIdentity"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tiao.session"
      }
    }
  }
}