# server/routes/game-auth.routes.ts
import { DeleteObjectCommand, PutObjectCommand } from "@aws-sdk/client-s3";
import bcrypt from "bcryptjs";
import { randomUUID } from "crypto";
import express, { Request, Response } from "express";
import { Jimp } from "jimp";
import mongoose from "mongoose";
import GameAccount from "../models/GameAccount";
import {
  clearPlayerSession,
  commitPlayerSession,
  createAccountAuth,
  createGuestAuth,
  deriveDisplayNameFromEmail,
  getPlayerFromRequest,
  refreshPlayerSession,
  sanitizeDisplayName,
} from "../game/playerTokens";
import { BUCKET_NAME, CLOUDFRONT_URL } from "../config/envVars";
import { s3Client } from "../config/s3Client";
import { multerUploadMiddleware } from "../middleware/multerUploadMiddleware";

const router = express.Router();
const saltRounds = 10;

/**
 * @openapi
 * components:
 *   schemas:
 *     PlayerIdentity:
 *       type: object
 *       properties:
 *         playerId:
 *           type: string
 *         displayName:
 *           type: string
 *         kind:
 *           type: string
 *           enum: [guest, account]
 *         email:
 *           type: string
 *         profilePicture:
 *           type: string
 *     AuthResponse:
 *       type: object
 *       properties:
 *         player:
 *           $ref: '#/components/schemas/PlayerIdentity'
 *     MultiplayerSnapshot:
 *       type: object
 *       properties:
 *         gameId:
 *           type: string
 *         roomType:
 *           type: string
 *           enum: [direct, matchmaking]
 *         status:
 *           type: string
 *           enum: [waiting, active, finished]
 *         state:
 *           type: object
 *         players:
 *           type: array
 *           items:
 *             type: object
 *         seats:
 *           type: object
 *         rematch:
 *           type: object
 *           nullable: true
 *     MatchmakingState:
 *       type: object
 *       properties:
 *         status:
 *           type: string
 *           enum: [idle, searching, matched]
 *     SocialOverview:
 *       type: object
 *       properties:
 *         friends:
 *           type: array
 *           items:
 *             type: object
 *         incomingFriendRequests:
 *           type: array
 *           items:
 *             type: object
 *         outgoingFriendRequests:
 *           type: array
 *           items:
 *             type: object
 *         incomingInvitations:
 *           type: array
 *           items:
 *             type: object
 *         outgoingInvitations:
 *           type: array
 *           items:
 *             type: object
 *   securitySchemes:
 *     sessionCookie:
 *       type: apiKey
 *       in: cookie
 *       name: tiao.session
 */

function isDatabaseReady(): boolean {
  return mongoose.connection.readyState === 1;
}

function buildAccountAuth(account: {
  id: string;
  email?: string;
  displayName: string;
  profilePicture?: string;
}) {
  return createAccountAuth({
    id: account.id,
    email: account.email,
    displayName: account.displayName,
    profilePicture: account.profilePicture,
  });
}

function serializeAccountProfile(account: {
  displayName: string;
  email?: string;
  profilePicture?: string;
  createdAt?: Date;
  updatedAt?: Date;
}) {
  return {
    displayName: account.displayName,
    email: account.email,
    profilePicture: account.profilePicture,
    createdAt: account.createdAt?.toISOString(),
    updatedAt: account.updatedAt?.toISOString(),
  };
}

async function requireAccount(req: Request, res: Response) {
  if (!isDatabaseReady()) {
    res.status(503).json({
      message:
        "Account features are unavailable right now. You can still keep playing as a guest.",
    });
    return null;
  }

  const player = await getPlayerFromRequest(req);
  if (!player) {
    res.status(401).json({
      message: "Not authenticated.",
    });
    return null;
  }

  if (player.kind !== "account") {
    res.status(403).json({
      message: "Only account players can edit a server profile.",
    });
    return null;
  }

  const account = await GameAccount.findById(player.playerId);
  if (!account) {
    res.status(404).json({
      message: "That account could not be found.",
    });
    return null;
  }

  return account;
}

/**
 * @openapi
 * /api/player/guest:
 *   post:
 *     summary: Create a guest session
 *     tags:
 *       - Authentication
 *     requestBody:
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               displayName:
 *                 type: string
 *     responses:
 *       201:
 *         description: Guest session created
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/AuthResponse'
 */
router.post("/guest", async (req: Request, res: Response) => {
  const { displayName } = req.body as {
    displayName?: string;
  };

  const auth = createGuestAuth(displayName);
  await commitPlayerSession(req, res, auth.player);
  res.status(201).json(auth);
});

/**
 * @openapi
 * /api/player/signup:
 *   post:
 *     summary: Create a new account
 *     tags:
 *       - Authentication
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               email:
 *                 type: string
 *                 format: email
 *               password:
 *                 type: string
 *                 minLength: 8
 *               displayName:
 *                 type: string
 *                 minLength: 3
 *     responses:
 *       201:
 *         description: Account created
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/AuthResponse'
 *       400:
 *         description: Validation error (missing fields, short password, short username, invalid email)
 *       409:
 *         description: Email or username already taken
 *       503:
 *         description: Account signup unavailable
 */
router.post("/signup", async (req: Request, res: Response) => {
  if (!isDatabaseReady()) {
    return res.status(503).json({
      message:
        "Account signup is unavailable right now. You can still keep playing as a guest.",
    });
  }

  const { email, password, displayName } = req.body as {
    email?: string;
    password?: string;
    displayName?: string;
  };

  const normalizedEmail = email?.trim().toLowerCase();
  const trimmedDisplayName = displayName?.trim();

  if (!password || (!normalizedEmail && !trimmedDisplayName)) {
    return res.status(400).json({
      message: "Provide a username or email address, and a password.",
    });
  }

  if (normalizedEmail) {
    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/;
    if (!emailRegex.test(normalizedEmail)) {
      return res.status(400).json({
        message: "Provide a valid email address.",
      });
    }

    const existingAccountByEmail = await GameAccount.findOne({
      email: normalizedEmail,
    });

    if (existingAccountByEmail) {
      return res.status(409).json({
        message: "An account with that email already exists.",
      });
    }
  }

  if (trimmedDisplayName) {
    if (trimmedDisplayName.length < 3) {
      return res.status(400).json({
        message: "Usernames must be at least 3 characters long.",
      });
    }

    const existingAccountByDisplayName = await GameAccount.findOne({
      displayName: trimmedDisplayName,
    });

    if (existingAccountByDisplayName) {
      return res.status(409).json({
        message: "That username is already taken.",
      });
    }
  }

  if (password.length < 8) {
    return res.status(400).json({
      message: "Passwords must be at least 8 characters long.",
    });
  }

  const passwordHash = bcrypt.hashSync(password, saltRounds);
  const account = await GameAccount.create({
    email: normalizedEmail || undefined,
    passwordHash,
    displayName: trimmedDisplayName || (normalizedEmail ? deriveDisplayNameFromEmail(normalizedEmail) : `Player-${randomUUID().slice(0, 8)}`),
  });

  const auth = buildAccountAuth(account);
  await commitPlayerSession(req, res, auth.player);
  return res.status(201).json(auth);
});

/**
 * @openapi
 * /api/player/login:
 *   post:
 *     summary: Log in to an existing account
 *     tags:
 *       - Authentication
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - identifier
 *               - password
 *             properties:
 *               identifier:
 *                 type: string
 *                 description: Username or email address
 *               password:
 *                 type: string
 *     responses:
 *       200:
 *         description: Login successful
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/AuthResponse'
 *       400:
 *         description: Missing identifier or password
 *       401:
 *         description: Account not found or incorrect password
 *       503:
 *         description: Account login unavailable
 */
router.post("/login", async (req: Request, res: Response) => {
  if (!isDatabaseReady()) {
    return res.status(503).json({
      message:
        "Account login is unavailable right now. You can still keep playing as a guest.",
    });
  }

  const { identifier, password } = req.body as {
    identifier?: string;
    password?: string;
  };

  if (!identifier || !password) {
    return res.status(400).json({
      message: "Provide a username or email address, and a password.",
    });
  }

  const trimmedIdentifier = identifier.trim();
  const lowercaseIdentifier = trimmedIdentifier.toLowerCase();

  const account = await GameAccount.findOne({
    $or: [
      { email: lowercaseIdentifier },
      { displayName: trimmedIdentifier },
    ],
  });

  if (!account) {
    return res.status(401).json({
      message: "No account was found with that identifier.",
    });
  }

  const passwordMatches = bcrypt.compareSync(password, account.passwordHash);
  if (!passwordMatches) {
    return res.status(401).json({
      message: "That password was incorrect.",
    });
  }

  const auth = buildAccountAuth(account);
  await commitPlayerSession(req, res, auth.player);
  return res.status(200).json(auth);
});

/**
 * @openapi
 * /api/player/logout:
 *   post:
 *     summary: Destroy the current session
 *     tags:
 *       - Authentication
 *     security:
 *       - sessionCookie: []
 *     responses:
 *       204:
 *         description: Session destroyed
 */
router.post("/logout", async (req: Request, res: Response) => {
  await clearPlayerSession(req, res);
  return res.status(204).send();
});

/**
 * @openapi
 * /api/player/me:
 *   get:
 *     summary: Get the current authenticated player
 *     tags:
 *       - Authentication
 *     security:
 *       - sessionCookie: []
 *     responses:
 *       200:
 *         description: Current player identity
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 player:
 *                   $ref: '#/components/schemas/PlayerIdentity'
 *       401:
 *         description: Not authenticated or session no longer valid
 */
router.get("/me", async (req: Request, res: Response) => {
  const player = await getPlayerFromRequest(req);
  if (!player) {
    return res.status(401).json({
      message: "Not authenticated.",
    });
  }

  if (player.kind === "account" && isDatabaseReady()) {
    const account = await GameAccount.findById(player.playerId);
    if (account) {
      const auth = buildAccountAuth(account);
      await refreshPlayerSession(req, res, auth.player);
      return res.status(200).json({
        player: auth.player,
      });
    }

    await clearPlayerSession(req, res);
    return res.status(401).json({
      message: "That account session is no longer valid.",
    });
  }

  await refreshPlayerSession(req, res, player);
  return res.status(200).json({ player });
});

/**
 * @openapi
 * /api/player/profile:
 *   get:
 *     summary: Get the current account profile
 *     tags:
 *       - Profile
 *     security:
 *       - sessionCookie: []
 *     responses:
 *       200:
 *         description: Account profile
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 profile:
 *                   type: object
 *                   properties:
 *                     displayName:
 *                       type: string
 *                     email:
 *                       type: string
 *                     profilePicture:
 *                       type: string
 *                     createdAt:
 *                       type: string
 *                       format: date-time
 *                     updatedAt:
 *                       type: string
 *                       format: date-time
 *       401:
 *         description: Not authenticated
 *       403:
 *         description: Only account players can access profiles
 *       404:
 *         description: Account not found
 *       503:
 *         description: Account features unavailable
 */
router.get("/profile", async (req: Request, res: Response) => {
  const account = await requireAccount(req, res);
  if (!account) {
    return;
  }

  return res.status(200).json({
    profile: serializeAccountProfile(account),
  });
});

/**
 * @openapi
 * /api/player/profile:
 *   put:
 *     summary: Update the current account profile
 *     tags:
 *       - Profile
 *     security:
 *       - sessionCookie: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               displayName:
 *                 type: string
 *                 minLength: 3
 *               email:
 *                 type: string
 *                 format: email
 *               password:
 *                 type: string
 *                 minLength: 8
 *     responses:
 *       200:
 *         description: Profile updated
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 auth:
 *                   $ref: '#/components/schemas/AuthResponse'
 *                 profile:
 *                   type: object
 *                   properties:
 *                     displayName:
 *                       type: string
 *                     email:
 *                       type: string
 *                     profilePicture:
 *                       type: string
 *                     createdAt:
 *                       type: string
 *                       format: date-time
 *                     updatedAt:
 *                       type: string
 *                       format: date-time
 *       400:
 *         description: Validation error (no fields provided, short display name, short password, invalid email)
 *       401:
 *         description: Not authenticated
 *       403:
 *         description: Only account players can edit profiles
 *       404:
 *         description: Account not found
 *       409:
 *         description: Username or email already taken
 *       503:
 *         description: Account features unavailable
 */
router.put("/profile", async (req: Request, res: Response) => {
  const account = await requireAccount(req, res);
  if (!account) {
    return;
  }

  const { displayName, email, password } = req.body as {
    displayName?: string;
    email?: string;
    password?: string;
  };

  const normalizedEmail = email?.trim().toLowerCase();
  const sanitizedDisplayName = displayName?.trim();

  if (!normalizedEmail && !sanitizedDisplayName && !password) {
    return res.status(400).json({
      message: "Provide a display name, email address, or password to update.",
    });
  }

  if (sanitizedDisplayName !== undefined) {
    if (!sanitizedDisplayName || sanitizedDisplayName.length < 3) {
      return res.status(400).json({
        message: "Display name must be at least 3 characters long.",
      });
    }

    const existingAccountByDisplayName = await GameAccount.findOne({
      displayName: sanitizedDisplayName,
      _id: { $ne: account._id },
    });

    if (existingAccountByDisplayName) {
      return res.status(409).json({
        message: "That username is already taken.",
      });
    }

    account.displayName = sanitizeDisplayName(sanitizedDisplayName);
  }

  if (normalizedEmail !== undefined) {
    if (normalizedEmail) {
      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/;
      if (!emailRegex.test(normalizedEmail)) {
        return res.status(400).json({
          message: "Provide a valid email address.",
        });
      }

      const existingAccount = await GameAccount.findOne({
        email: normalizedEmail,
        _id: { $ne: account._id },
      });

      if (existingAccount) {
        return res.status(409).json({
          message: "An account with that email already exists.",
        });
      }

      account.email = normalizedEmail;
    } else {
      account.email = undefined;
    }
  }

  if (password !== undefined) {
    if (password.length < 8) {
      return res.status(400).json({
        message: "Passwords must be at least 8 characters long.",
      });
    }

    account.passwordHash = bcrypt.hashSync(password, saltRounds);
  }

  await account.save();

  const auth = buildAccountAuth(account);
  await refreshPlayerSession(req, res, auth.player);
  return res.status(200).json({
    auth,
    profile: serializeAccountProfile(account),
  });
});

/**
 * @openapi
 * /api/player/profile-picture:
 *   post:
 *     summary: Upload a profile picture
 *     tags:
 *       - Profile
 *     security:
 *       - sessionCookie: []
 *     requestBody:
 *       required: true
 *       content:
 *         multipart/form-data:
 *           schema:
 *             type: object
 *             required:
 *               - profilePicture
 *             properties:
 *               profilePicture:
 *                 type: string
 *                 format: binary
 *     responses:
 *       200:
 *         description: Profile picture uploaded
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 auth:
 *                   $ref: '#/components/schemas/AuthResponse'
 *                 profile:
 *                   type: object
 *                   properties:
 *                     displayName:
 *                       type: string
 *                     email:
 *                       type: string
 *                     profilePicture:
 *                       type: string
 *                     createdAt:
 *                       type: string
 *                       format: date-time
 *                     updatedAt:
 *                       type: string
 *                       format: date-time
 *       400:
 *         description: No image file provided
 *       401:
 *         description: Not authenticated
 *       403:
 *         description: Only account players can upload profile pictures
 *       404:
 *         description: Account not found
 *       500:
 *         description: Upload failed
 *       503:
 *         description: Account features unavailable
 */
router.post(
  "/profile-picture",
  multerUploadMiddleware.single("profilePicture"),
  async (req: Request, res: Response) => {
    const account = await requireAccount(req, res);
    if (!account) {
      return;
    }

    if (!req.file) {
      return res.status(400).json({
        message: "Choose an image to upload.",
      });
    }

    try {
      const fileName = `game-account-${account.id}-${randomUUID()}.jpeg`;
      const image = await Jimp.read(req.file.buffer);
      image.resize({ w: 320 });

      const processedImageBuffer = await image.getBuffer("image/jpeg");

      const uploadCommand = new PutObjectCommand({
        Bucket: BUCKET_NAME,
        Key: fileName,
        Body: processedImageBuffer,
        ContentType: "image/jpeg",
      });

      await s3Client.send(uploadCommand);

      if (account.profilePicture?.startsWith(`${CLOUDFRONT_URL}/`)) {
        try {
          const previousKey = account.profilePicture.split("/").pop();
          if (previousKey) {
            await s3Client.send(
              new DeleteObjectCommand({
                Bucket: BUCKET_NAME,
                Key: previousKey,
              })
            );
          }
        } catch (error) {
          console.error("Error deleting previous game account profile picture:", error);
        }
      }

      account.profilePicture = `${CLOUDFRONT_URL}/${fileName}`;
      await account.save();

      const auth = buildAccountAuth(account);
      await refreshPlayerSession(req, res, auth.player);
      return res.status(200).json({
        auth,
        profile: serializeAccountProfile(account),
      });
    } catch (error) {
      console.error("Error uploading game account profile picture:", error);
      return res.status(500).json({
        message: "Unable to upload that profile picture right now.",
      });
    }
  }
);

export default router;